+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file's format is based on [Keep a Changelog](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Pending
+### Fixed
+ - make tax request throw correct exception when no line items (!34)
+
 ## 3.1.0 - 2018-09-18
 ### Added
  - Adds /shop/api/place-order endpoint (!32)